add Subscribers.wrap
