+++ src/main/java/rx/internal/operators/OnSubscribeDelaySubscription.java
@@ -21,6 +21,7 @@ import rx.*;
 import rx.Observable.OnSubscribe;
 import rx.Scheduler.Worker;
 import rx.functions.Action0;
+import rx.observers.Subscribers;
 
 /**
  * Delays the subscription to the source by the given amount, running on the given scheduler.
@@ -49,20 +50,7 @@ public final class OnSubscribeDelaySubscription<T> implements OnSubscribe<T> {
             @Override
             public void call() {
                 if (!s.isUnsubscribed()) {
-                    source.unsafeSubscribe(new Subscriber<T>(s) {
-                        @Override
-                        public void onNext(T t) {
-                            s.onNext(t);
-                        }
-                        @Override
-                        public void onError(Throwable e) {
-                            s.onError(e);
-                        }
-                        @Override
-                        public void onCompleted() {
-                            s.onCompleted();
-                        }
-                    });
+                    source.unsafeSubscribe(Subscribers.wrap(s));
                 }
             }
         }, time, unit);
